change hton{s,l} to htobe{16,32,64} and toh{s,l} change set recvbuf after makeSocket change MAX_BLOCK_SIZE to 1472-sizeof(serverDataMsg)

+++ dyode-half-fiber/udpcast/fifo.c
@@ -1,7 +1,7 @@
 #include "libbb_udpcast.h"
 #include "fifo.h"
 
-void udpc_initFifo(struct fifo *fifo, unsigned int blockSize)
+void udpc_initFifo(struct fifo *fifo, size_t blockSize)
 {
     fifo->dataBufSize = blockSize * 4096;
     fifo->dataBuffer = xmalloc(fifo->dataBufSize+4096);
