+++ dyode-half-fiber/udpcast/participants.h
@@ -18,13 +18,13 @@
 #define printSet udpc_printSet
 
 struct  participantsDb {
-    int32_t nrParticipants;
+    int16_t nrParticipants;
 
     struct clientDesc {
         struct sockaddr_in addr;
         int used;
-        int capabilities;
-        unsigned int rcvbuf;
+        uint32_t capabilities;
+        uint32_t rcvbuf;
     } clientTable[MAX_CLIENTS];
 };
 
@@ -32,17 +32,17 @@ typedef struct participantsDb *participantsDb_t;
 
 int udpc_isParticipantValid(participantsDb_t ,int32_t slot);
 int udpc_removeParticipant(participantsDb_t, int32_t slot);
-int32_t udpc_lookupParticipant(participantsDb_t, struct sockaddr_in *addr);
-int32_t udpc_nrParticipants(participantsDb_t);
-int32_t udpc_addParticipant(participantsDb_t,
+int16_t udpc_lookupParticipant(participantsDb_t, struct sockaddr_in *addr);
+int16_t udpc_nrParticipants(participantsDb_t);
+int16_t udpc_addParticipant(participantsDb_t,
                              struct sockaddr_in *addr,
-                             int capabilities,
-                             unsigned int rcvbuf,
+                             uint32_t capabilities,
+                             uint32_t rcvbuf,
                              int pointopoint);
 participantsDb_t udpc_makeParticipantsDb(void);
-int udpc_getParticipantCapabilities(participantsDb_t db, int32_t i);
-unsigned int udpc_getParticipantRcvBuf(participantsDb_t db, int32_t i);
-struct sockaddr_in *udpc_getParticipantIp(participantsDb_t db, int32_t i);
+uint32_t udpc_getParticipantCapabilities(participantsDb_t db, int16_t i);
+uint32_t udpc_getParticipantRcvBuf(participantsDb_t db, int16_t i);
+struct sockaddr_in *udpc_getParticipantIp(participantsDb_t db, int16_t i);
 void udpc_printNotSet(participantsDb_t db, char *d);
 void udpc_printSet(participantsDb_t db, char *d);
